# Improve the memory usage and performance of the cache plugin