### 1. Why is this pull request needed and what does it do?
This PR reduces the memory usage of the cache plugin (14 vs 8 allocs and 540 vs 327 B) and improves performance as a side effect.

**PR:**
```
goos: darwin
goarch: amd64
pkg: github.com/coredns/coredns/plugin/cache
BenchmarkCacheResponse/NoPrefetch-12         	 1601374	       732 ns/op	     327 B/op	       8 allocs/op
BenchmarkCacheResponse/Prefetch-12           	 1434120	       830 ns/op	     327 B/op	       8 allocs/op
BenchmarkHash-12                             	130255627	         9.25 ns/op	       0 B/op	       0 allocs/op
PASS
```

master@ 004c5fca9d7dfb8b58d608e4d50833b40953441f
```
goos: darwin
goarch: amd64
pkg: github.com/coredns/coredns/plugin/cache
BenchmarkCacheResponse-12    	 1181955	       971 ns/op	     540 B/op	      14 allocs/op
PASS
```

### 2. Which issues (if any) are related?
N/A
### 3. Which documentation changes (if any) need to be made?
N/A
### 4. Does this introduce a backward incompatible change or deprecation?
N/A