@miekg 2019-09-23
Note there is also this closed PR https://github.com/coredns/coredns/pull/2527/files that tackles this on a more fundamental level, but not storing parts of a message that don't need to be stored. That didn't get merged because getting actual memory improvement data was hard.