@miekg 2019-09-24
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] Improve the m..." ]
>charlievieth commented on this pull request.
>
>
>
>> @@ -76,21 +75,39 @@ func key(qname string, m *dns.Msg, t response.Type, do bool) (bool, uint64) {
> 	return true, hash(qname, m.Question[0].Qtype, do)
> }
>
>-var one = []byte("1")
>-var zero = []byte("0")
>+type fnv64 uint64
>
>I'll document it, this is the same `fnv` function as the stdlib, but has a WriteByte and WriteString method.

I'm not a fan of rewriting parts of the stdlib, we'll now miss out any fixes 
(like ASM) they apply to this. Is there a smarter way of doing this? Dropping 
FNV all together?