@miekg 2019-10-03
I'm moving some of this into their own PRs. For one is prefetch closure, see: #3353